优化页面。添加控制项

+++ test/src/framework/modules/lowvalueitemsmanagement/bo/CategoryManagementBO.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.UUID;
 
 import framework.modules.lowvalueitemsmanagement.dao.CategoryManagementDAO;
+import framework.modules.lowvalueitemsmanagement.dao.ItemManageDAO;
 import framework.modules.lowvalueitemsmanagement.domain.CategoryManagement;
 import framework.sys.basemodule.bo.BOBase;
 import framework.sys.context.applicationworker.MethodID;
@@ -13,9 +14,15 @@ import framework.sys.tools.DBOperation;
 @LogOperate(menu = "低值易耗品类目管理")
 public class CategoryManagementBO extends BOBase<CategoryManagementDAO, CategoryManagement>{
 
+	private ItemManageDAO itemManageDAO;
+	
 	@MethodID("addCategory")
 	@LogOperate(operate = "新增类目")
-	public void addCategory_log_trans(CategoryManagement categoryManagement) {
+	public String addCategory_log_trans(CategoryManagement categoryManagement) {
+		boolean flag = entityDAO.executeFindExists("select 1 from tCategoryManagement where categoryName = ?", categoryManagement.getCategoryName());
+		if (flag) {
+			return "类目名称已经存在，请重新输入";
+		}
 		String pk = UUID.randomUUID().toString();
 		categoryManagement.setPk(pk);
 		String[] updateInfo = DBOperation.getUpdateInfo();
@@ -23,6 +30,8 @@ public class CategoryManagementBO extends BOBase<CategoryManagementDAO, Category
 		categoryManagement.setLastestUpdate(updateInfo[0]);
 		categoryManagement.setUpdatePerson(updateInfo[2]);
 		entityDAO.save(categoryManagement);
+		
+		return "";
 	}
 	
 	@MethodID("modifyCategory")
@@ -37,9 +46,12 @@ public class CategoryManagementBO extends BOBase<CategoryManagementDAO, Category
 	@MethodID("deleteCategory")
 	@LogOperate(operate = "删除一条类目")
 	public String deleteCategory_log_trans(String pk) {
-		String return_tips = "";
+		boolean isExistItem = itemManageDAO.executeFindExists("select 1 from titemManage where imCategoryPK=?", pk);
+		if (isExistItem) {
+			return "类目在物品管理表中有记录时，不允许删除";
+		}
 		entityDAO.delete(entityDAO.findById(pk));
-		return return_tips;
+		return "";
 	}
 	@MethodID("findCategoryByGroupCode")
 	public List<CategoryManagement> findCategoryByGroupCode(String groupCode) {
@@ -63,5 +75,13 @@ public class CategoryManagementBO extends BOBase<CategoryManagementDAO, Category
 
 		*//** 第二步：删除类目对应的各种信息 * *//*
 	}*/
+
+	public ItemManageDAO getItemManageDAO() {
+		return itemManageDAO;
+	}
+
+	public void setItemManageDAO(ItemManageDAO itemManageDAO) {
+		this.itemManageDAO = itemManageDAO;
+	}
 	
 }
